+++ color-manager/src/test/java/com/my_wall_color/color_manager/color/adapter/jpa/JpaColorMapperTest.java
@@ -11,7 +11,7 @@ class JpaColorMapperTest {
 
     @Test
     void fromDomain() {
-        Color color = Color.create(971, (short) 10, (short) 20, (short) 30, "test-color", 31);
+        Color color = Color.create(971, 10, 20, 30, "test-color", 31);
         JpaColor jpaColor = unitUnderTest.fromDomain(color);
         assertThat(jpaColor.getId()).isEqualTo(971);
         assertThat(jpaColor.getRed()).isEqualTo((short) 10);
