+++ color-manager/src/test/java/com/my_wall_color/color_manager/color/domain/ColorTest.java
@@ -8,21 +8,21 @@ import static org.junit.jupiter.api.Assertions.*;
 class ColorTest {
     @Test
     void shouldNotAcceptInvalidColorComponent() {
-        assertThrows(IllegalArgumentException.class, () -> Color.create(null, (short) 255, (short) 255, Short.MIN_VALUE, "min blue", 1));
-        assertThrows(IllegalArgumentException.class, () -> Color.create(null, (short) 255, (short) 256, (short) 0, "green above 1", 1));
-        assertThrows(IllegalArgumentException.class, () -> Color.create(null, Short.MAX_VALUE, (short) 255, (short) 0, "max red", 1));
-        assertThrows(IllegalArgumentException.class, () -> Color.create(null, (short) -1, (short) 255, (short) 0, "red below 1", 1));
+        assertThrows(IllegalArgumentException.class, () -> Color.create(null, 255, 255, Short.MIN_VALUE, "min blue", 1));
+        assertThrows(IllegalArgumentException.class, () -> Color.create(null, 255, 256, 0, "green above 1", 1));
+        assertThrows(IllegalArgumentException.class, () -> Color.create(null, Short.MAX_VALUE, 255, 0, "max red", 1));
+        assertThrows(IllegalArgumentException.class, () -> Color.create(null, -1, 255, 0, "red below 1", 1));
     }
 
     @Test
     void shouldNotAcceptEmptyName() {
-        assertThrows(IllegalArgumentException.class, () -> Color.create(null, (short) 255, (short) 255, (short) 0, "", 1));
-        assertThrows(IllegalArgumentException.class, () -> Color.create(null, (short) 255, (short) 255, (short) 0, null, 1));
+        assertThrows(IllegalArgumentException.class, () -> Color.create(null, 255, 255, 0, "", 1));
+        assertThrows(IllegalArgumentException.class, () -> Color.create(null, 255, 255, 0, null, 1));
     }
 
     @Test
     void shouldCreateColor() {
-        var color = Color.create(null, (short) 200, (short) 100, (short) 50, "random color", 5);
+        var color = Color.create(null, 200, 100, 50, "random color", 5);
         assertThat(color.getId()).isNull();
         assertThat(color.getRed()).isEqualTo((short) 200);
         assertThat(color.getGreen()).isEqualTo((short) 100);
